Tests for dynamic map update. Added sh script to extract pngs from videos.

+++ tools/generate_results_amcl.sh
@@ -5,7 +5,7 @@
 ##################################################
 
 results_directory=${1:?'Must specify directory where the results_amcl.bag is and in which the results will be saved'}
-
+extract_cvs_from_bag=${2:-true}
 
 echo "############################################################################################################################################################"
 echo "##### Generating results for path: ${results_directory}"
@@ -13,27 +13,33 @@ echo "##########################################################################
 
 
 # generate CSVs from rosbag
-mkdir -p "${results_directory}/amcl_bag"
-mv "${results_directory}/results_amcl.bag" "${results_directory}/amcl_bag/results_amcl.bag"
-rosrun robot_localization_tools bag2csv.sh "${results_directory}/amcl_bag/results_amcl" '/dynamic_robot_localization/localization_error /dynamic_robot_localization/odometry_error /rosout /amcl_pose'
+if [ "${extract_cvs_from_bag}" = true ]; then
+	mkdir -p "${results_directory}/amcl_bag"
+	mv "${results_directory}/results_amcl.bag" "${results_directory}/amcl_bag/results_amcl.bag"
+	rosrun robot_localization_tools bag2csv.sh "${results_directory}/amcl_bag/results_amcl" '/dynamic_robot_localization/localization_error /dynamic_robot_localization/odometry_error /rosout /amcl_pose'
+	
+	mv "${results_directory}/amcl_bag/results_amcl.bag" "${results_directory}/results_amcl.bag"
+	mv "${results_directory}/amcl_bag/results_amcl__dynamic_robot_localization_localization_error.csv" "${results_directory}/results__amcl_localization_error.csv"
+	mv "${results_directory}/amcl_bag/results_amcl__dynamic_robot_localization_odometry_error.csv" "${results_directory}/results__amcl_odometry_error.csv"
+	mv "${results_directory}/amcl_bag/results_amcl__rosout.csv" "${results_directory}/results__rosout_amcl.csv"
+	mv "${results_directory}/amcl_bag/results_amcl__amcl_pose.csv" "${results_directory}/results__amcl_pose.csv"
+	
+	rm -rf "${results_directory}/amcl_bag"
+	
+	mkdir -p "${results_directory}/pdf"
+	mkdir -p "${results_directory}/svg"
+	mkdir -p "${results_directory}/eps"
+fi
 
-mv "${results_directory}/amcl_bag/results_amcl.bag" "${results_directory}/results_amcl.bag"
-mv "${results_directory}/amcl_bag/results_amcl__dynamic_robot_localization_localization_error.csv" "${results_directory}/results__amcl_localization_error.csv"
-mv "${results_directory}/amcl_bag/results_amcl__dynamic_robot_localization_odometry_error.csv" "${results_directory}/results__amcl_odometry_error.csv"
-mv "${results_directory}/amcl_bag/results_amcl__rosout.csv" "${results_directory}/results__rosout_amcl.csv"
-mv "${results_directory}/amcl_bag/results_amcl__amcl_pose.csv" "${results_directory}/results__amcl_pose.csv"
 
-rm -rf "${results_directory}/amcl_bag"
-
-mkdir -p "${results_directory}/pdf"
-mkdir -p "${results_directory}/svg"
-mkdir -p "${results_directory}/eps"
-
-path_files="${results_directory}/results_ground_truth_poses.txt+${results_directory}/results_localization_poses.txt+${results_directory}/results_localization_poses_amcl.txt+${results_directory}/results_odometry_poses.txt"
 
 echo "\n======================================================================================="
 echo "Building path (with arrows) from the ground truth and localization system poses"
-rosrun robot_localization_tools path_plotter.py -i ${path_files} -o ${results_directory}/robot-movement-path-with-odometry-and-amcl -p 1 -v 8 -a 0.0025 -c 'g+b+c+r' -t 'Robot movement path (green -> ground truth, blue -> localization system, cyan -> amcl, red -> odometry)' -s 1 -q 1 -d 0 &
+
+graphs_colors='g+b+#5C3317+r'
+
+path_files="${results_directory}/results_ground_truth_poses.txt+${results_directory}/results_localization_poses.txt+${results_directory}/results_localization_poses_amcl.txt+${results_directory}/results_odometry_poses.txt"
+rosrun robot_localization_tools path_plotter.py -i ${path_files} -o ${results_directory}/robot-movement-path-with-odometry-and-amcl -p 1 -v 8 -a 0.005 -c ${graphs_colors} -t 'Robot movement path (green -> ground truth, blue -> localization system, brown -> amcl, red -> odometry)' -s 1 -q 1 -d 0 &
 
 
 echo "\n======================================================================================="
@@ -43,23 +49,23 @@ graphs_common_configs="-k 0.75 -r 1 -g 1 -s 1 -q 1 -d 0"
 path_graphs_common_configs="--reset 1 --grid 1 -k 0.75 -r 10 -g 3 -s 1 -q 1 -d 0"
 
 # Evolution of x, y and z
-rosrun robot_localization_tools graph_plotter.py -i ${path_files} -o ${results_directory}/robot-movement-path-position-evolution-x-with-amcl -x '0-0-0-0' -y '1-1-1-1' -z ' ' -e 2 -w 0.25 -m 1 -n 1 -b 'Time of localization update (seconds from beginning of test)' -v 'X position (meters)' -l 'Ground truth x positions+Localization system x positions+AMCL+Odometry x positions' -c 'g+b+c+r' -t 'Evolution of x position' ${graphs_common_configs} &
-rosrun robot_localization_tools graph_plotter.py -i ${path_files} -o ${results_directory}/robot-movement-path-position-evolution-y-with-amcl -x '0-0-0-0' -y '2-2-2-2' -z ' ' -e 2 -w 0.25 -m 1 -n 1 -b 'Time of localization update (seconds from beginning of test)' -v 'Y position (meters)' -l 'Ground truth y positions+Localization system y positions+AMCL+Odometry y positions' -c 'g+b+c+r' -t 'Evolution of y position' ${graphs_common_configs} &
-rosrun robot_localization_tools graph_plotter.py -i ${path_files} -o ${results_directory}/robot-movement-path-position-evolution-z-with-amcl -x '0-0-0-0' -y '3-3-3-3' -z ' ' -e 2 -w 0.25 -m 1 -n 1 -b 'Time of localization update (seconds from beginning of test)' -v 'Z position (meters)' -l 'Ground truth z positions+Localization system z positions+AMCL+Odometry z positions' -c 'g+b+c+r' -t 'Evolution of z position' ${graphs_common_configs} &
+rosrun robot_localization_tools graph_plotter.py -i ${path_files} -o ${results_directory}/robot-movement-path-position-evolution-x-with-amcl -x '0-0-0-0' -y '1-1-1-1' -z ' ' -e 2 -w 0.25 -m 1 -n 1 -b 'Time of localization update (seconds from beginning of test)' -v 'X position (meters)' -l 'Ground truth+Localization system+AMCL+Odometry' -c ${graphs_colors} -t 'Evolution of x position' ${graphs_common_configs} &
+rosrun robot_localization_tools graph_plotter.py -i ${path_files} -o ${results_directory}/robot-movement-path-position-evolution-y-with-amcl -x '0-0-0-0' -y '2-2-2-2' -z ' ' -e 2 -w 0.25 -m 1 -n 1 -b 'Time of localization update (seconds from beginning of test)' -v 'Y position (meters)' -l 'Ground truth+Localization system+AMCL+Odometry' -c ${graphs_colors} -t 'Evolution of y position' ${graphs_common_configs} &
+rosrun robot_localization_tools graph_plotter.py -i ${path_files} -o ${results_directory}/robot-movement-path-position-evolution-z-with-amcl -x '0-0-0-0' -y '3-3-3-3' -z ' ' -e 2 -w 0.25 -m 1 -n 1 -b 'Time of localization update (seconds from beginning of test)' -v 'Z position (meters)' -l 'Ground truth+Localization system+AMCL+Odometry' -c ${graphs_colors} -t 'Evolution of z position' ${graphs_common_configs} &
 
 # Cumulative position evolution
-rosrun robot_localization_tools path_velocity_and_acceleration_plotter.py -i ${path_files} -f -1 -o ${results_directory}/robot-movement-path-position-differences-with-amcl -p 1 -x '0-0-0-0' -y '1-1-1-1' -z ' ' -e 2 -w 0.25 -m 1 -n 1 -b 'Time of localization update (seconds from beginning of test)' -v 'Position differences from consecutive poses (meters)' -l 'Ground truth+Localization system+AMCL+Odometry' -c 'g+b+c+r' -t 'Position differences from consecutive poses' ${path_graphs_common_configs} &
-rosrun robot_localization_tools path_velocity_and_acceleration_plotter.py -i ${path_files} -f -1 -o ${results_directory}/robot-movement-path-angular-differences-with-amcl -p 0 -x '0-0-0-0' -y '4-4-4-4' -z ' ' -e 2 -w 0.25 -m 1 -n 1 -b 'Time of localization update (seconds from beginning of test)' -v 'Angular differences from consecutive poses (degrees)' -l 'Ground truth+Localization system+AMCL+Odometry' -c 'g+b+c+r' -t 'Angular differences from consecutive poses' ${path_graphs_common_configs} &
-rosrun robot_localization_tools path_velocity_and_acceleration_plotter.py -i ${path_files} -f 0 -o ${results_directory}/robot-movement-path-position-cumulative-with-amcl -p 1 -x '0-0-0-0' -y '1-1-1-1' -z ' ' -e 2 -w 0.25 -m 1 -n 1 -b 'Time of localization update (seconds from beginning of test)' -v 'Cumulative position differences from consecutive poses (meters)' -l 'Ground truth+Localization system+AMCL+Odometry' -c 'g+b+c+r' -t 'Cumulative position differences from consecutive poses' ${path_graphs_common_configs} &
-rosrun robot_localization_tools path_velocity_and_acceleration_plotter.py -i ${path_files} -f 0 -o ${results_directory}/robot-movement-path-angular-cumulative-with-amcl -p 0 -x '0-0-0-0' -y '4-4-4-4' -z ' ' -e 2 -w 0.25 -m 1 -n 1 -b 'Time of localization update (seconds from beginning of test)' -v 'Cumulative angular differences from consecutive poses (degrees)' -l 'Ground truth+Localization system+AMCL+Odometry' -c 'g+b+c+r' -t 'Cumulative angular differences from consecutive poses' ${path_graphs_common_configs} &
+rosrun robot_localization_tools path_velocity_and_acceleration_plotter.py -i ${path_files} -f -1 -o ${results_directory}/robot-movement-path-position-differences-with-amcl -p 1 -x '0-0-0-0' -y '1-1-1-1' -z ' ' -e 2 -w 0.25 -m 1 -n 1 -b 'Time of localization update (seconds from beginning of test)' -v 'Position differences from consecutive poses (meters)' -l 'Ground truth+Localization system+AMCL+Odometry' -c ${graphs_colors} -t 'Position differences from consecutive poses' ${path_graphs_common_configs} &
+rosrun robot_localization_tools path_velocity_and_acceleration_plotter.py -i ${path_files} -f -1 -o ${results_directory}/robot-movement-path-angular-differences-with-amcl -p 0 -x '0-0-0-0' -y '4-4-4-4' -z ' ' -e 2 -w 0.25 -m 1 -n 1 -b 'Time of localization update (seconds from beginning of test)' -v 'Angular differences from consecutive poses (degrees)' -l 'Ground truth+Localization system+AMCL+Odometry' -c ${graphs_colors} -t 'Angular differences from consecutive poses' ${path_graphs_common_configs} &
+rosrun robot_localization_tools path_velocity_and_acceleration_plotter.py -i ${path_files} -f 0 -o ${results_directory}/robot-movement-path-position-cumulative-with-amcl -p 1 -x '0-0-0-0' -y '1-1-1-1' -z ' ' -e 2 -w 0.25 -m 1 -n 1 -b 'Time of localization update (seconds from beginning of test)' -v 'Cumulative position differences from consecutive poses (meters)' -l 'Ground truth+Localization system+AMCL+Odometry' -c ${graphs_colors} -t 'Cumulative position differences from consecutive poses' ${path_graphs_common_configs} &
+rosrun robot_localization_tools path_velocity_and_acceleration_plotter.py -i ${path_files} -f 0 -o ${results_directory}/robot-movement-path-angular-cumulative-with-amcl -p 0 -x '0-0-0-0' -y '4-4-4-4' -z ' ' -e 2 -w 0.25 -m 1 -n 1 -b 'Time of localization update (seconds from beginning of test)' -v 'Cumulative angular differences from consecutive poses (degrees)' -l 'Ground truth+Localization system+AMCL+Odometry' -c ${graphs_colors} -t 'Cumulative angular differences from consecutive poses' ${path_graphs_common_configs} &
 
 # Velocity
-rosrun robot_localization_tools path_velocity_and_acceleration_plotter.py -i ${path_files} -f 1 -o ${results_directory}/robot-movement-path-linear-velocity-with-amcl -p 1 -x '0-0-0-0' -y '1-1-1-1' -z ' ' -e 2 -w 0.25 -m 1 -n 1 -b 'Time of localization update (seconds from beginning of test)' -v 'Linear velocity (meters / second)' -l 'Ground truth+Localization system+AMCL+Odometry' -c 'g+b+c+r' -t 'Estimated linear velocity' ${path_graphs_common_configs} &
-rosrun robot_localization_tools path_velocity_and_acceleration_plotter.py -i ${path_files} -f 1 -o ${results_directory}/robot-movement-path-angular-velocity-with-amcl -p 0 -x '0-0-0-0' -y '4-4-4-4' -z ' ' -e 2 -w 0.25 -m 1 -n 1 -b 'Time of localization update (seconds from beginning of test)' -v 'Angular velocity (degrees / second)' -l 'Ground truth+Localization system+AMCL+Odometry' -c 'g+b+c+r' -t 'Estimated angular velocity' ${path_graphs_common_configs} &
+rosrun robot_localization_tools path_velocity_and_acceleration_plotter.py -i ${path_files} -f 1 -o ${results_directory}/robot-movement-path-linear-velocity-with-amcl -p 1 -x '0-0-0-0' -y '1-1-1-1' -z ' ' -e 2 -w 0.25 -m 1 -n 1 -b 'Time of localization update (seconds from beginning of test)' -v 'Linear velocity (meters / second)' -l 'Ground truth+Localization system+AMCL+Odometry' -c ${graphs_colors} -t 'Estimated linear velocity' ${path_graphs_common_configs} &
+rosrun robot_localization_tools path_velocity_and_acceleration_plotter.py -i ${path_files} -f 1 -o ${results_directory}/robot-movement-path-angular-velocity-with-amcl -p 0 -x '0-0-0-0' -y '4-4-4-4' -z ' ' -e 2 -w 0.25 -m 1 -n 1 -b 'Time of localization update (seconds from beginning of test)' -v 'Angular velocity (degrees / second)' -l 'Ground truth+Localization system+AMCL+Odometry' -c ${graphs_colors} -t 'Estimated angular velocity' ${path_graphs_common_configs} &
 
 # Acceleration
-rosrun robot_localization_tools path_velocity_and_acceleration_plotter.py -i ${path_files} -f 2 -o ${results_directory}/robot-movement-path-linear-acceleration-with-amcl -p 1 -x '0-0-0-0' -y '1-1-1-1' -z ' ' -e 2 -w 0.25 -m 1 -n 1 -b 'Time of localization update (seconds from beginning of test)' -v 'Linear acceleration (meters / second / second)' -l 'Ground truth+Localization system+AMCL+Odometry' -c 'g+b+c+r' -t 'Estimated linear acceleration' ${path_graphs_common_configs} &
-rosrun robot_localization_tools path_velocity_and_acceleration_plotter.py -i ${path_files} -f 2 -o ${results_directory}/robot-movement-path-angular-acceleration-with-amcl -p 0 -x '0-0-0-0' -y '4-4-4-4' -z ' ' -e 2 -w 0.25 -m 1 -n 1 -b 'Time of localization update (seconds from beginning of test)' -v 'Angular acceleration (degrees / second / second)' -l 'Ground truth+Localization system+AMCL+Odometry' -c 'g+b+c+r' -t 'Estimated angular acceleration' ${path_graphs_common_configs} &
+rosrun robot_localization_tools path_velocity_and_acceleration_plotter.py -i ${path_files} -f 2 -o ${results_directory}/robot-movement-path-linear-acceleration-with-amcl -p 1 -x '0-0-0-0' -y '1-1-1-1' -z ' ' -e 2 -w 0.25 -m 1 -n 1 -b 'Time of localization update (seconds from beginning of test)' -v 'Linear acceleration (meters / second / second)' -l 'Ground truth+Localization system+AMCL+Odometry' -c ${graphs_colors} -t 'Estimated linear acceleration' ${path_graphs_common_configs} &
+rosrun robot_localization_tools path_velocity_and_acceleration_plotter.py -i ${path_files} -f 2 -o ${results_directory}/robot-movement-path-angular-acceleration-with-amcl -p 0 -x '0-0-0-0' -y '4-4-4-4' -z ' ' -e 2 -w 0.25 -m 1 -n 1 -b 'Time of localization update (seconds from beginning of test)' -v 'Angular acceleration (degrees / second / second)' -l 'Ground truth+Localization system+AMCL+Odometry' -c ${graphs_colors} -t 'Estimated angular acceleration' ${path_graphs_common_configs} &
 
 # Localization errors
 rosrun robot_localization_tools graph_plotter.py -i ${results_directory}/results__amcl_localization_error.csv -o ${results_directory}/translation-error-components-millimeters-amcl -x 2 -y '4+5+6' -w 0.25 -m 0.000000001 -n 1 -b 'Time of localization update (seconds from beginning of test)' -v 'Translation error (millimeters)' -l 'Translation error in the x axis+Translation error in the y axis+Translation error in the z axis' -c 'y+g+b' -t 'Translation errors by axis' ${graphs_common_configs} &
